+++ btclib/tests/test_curvegroupf.py
@@ -1,4 +1,3 @@
-
 #!/usr/bin/env python3
 
 # Copyright (C) 2020 The btclib developers